Move component handler execution to ComponentHandlerExecutor
